Added a lot, wrote more in the Description and commented allot
First I made the goal_reached function, in the board class, into a simple bool. Then I changed 'player_list' to 'player_field' so it passes the test. Then I removed the if function to the self.player_list, and made the self.board a into a if statement. I then begun on the single_game function. It's kind of right, but does not pass the test. But by making the single_game function workable, it was kind of easy to do the rest of the functions, so I finished all of them.  They all should work, but the returns just looks a bit weird, so we have to change that.